Relocating the test.
